services/catalog/requirements.txt

fastapi==0.116.0
uvicorn==0.35.0
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
SQLAlchemy==2.0.41
python-jose[cryptography]==3.5.0
python-multipart==0.0.6
python-dotenv==1.1.1
alembic==1.16.4
email-validator==2.1.1     # (pydantic extra)
uvloop==0.21.0
httptools==0.6.4
psycopg2-binary==2.9.9 